requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
cffi==1.12.3
cryptography==2.7
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
entrypoints==0.3
Faker==0.9.1
filelock==3.0.12
flake8==3.7.8
importlib-metadata==0.18
Markdown==3.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==7.2.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.0
tox==3.13.2
virtualenv==16.6.2
wcwidth==0.1.7
whitenoise==4.1.3
zipp==0.5.2
